Moved sleep shim to different class
